Display both coverage and press on homepage.
Unfortunately this requires adding a new field "sortDate" to all existing press releases. This is necessary because the existing "date" field is parsed by Hugo into a date object, which cannot be compared to the dates in the .yml file which are just parsed as strings. sortDate is always parsed into string, so it can be compared (and thus, sorted).